Refine setup.py (#10577)
* fix

* fix
